requirements.txt
pandas
numpy
matplotlib
joblib
seaborn
tqdm
scipy
dagshub
python-box==6.0.2
ensure==1.0.2
pyYAML
types-pyYAML
scipy
scikit-learn
category_encoder
-e .